#5 Stack구조체 분리
